Remove ability to serialize Task fields
Tasks are *not* serializable. And they should not be made serializable. Who knows when or if they'll complete, and what thread they need to be available to complete, etc.

If a class really wants to serialize a value that comes from a task, they can write the code to expose the Task's value as a property getter and return the value. And hopefully they can do so in a safe manner that will not deadlock.

Fixes #29