Get MessagePack working on UWP + iOS
With this change MessagePack builds specially for iOS and UWP applications and omits the Ref.Emit code similar to how we omit it for certain Unity platforms.

A UWP test project is added to verify that the most basic scenario works. We may add more/all of our unit tests to run in the UWP test project eventually.