Touch-up the comments and optimizations in MessagePackBinary
Among the optimizations, I replace an optimization to avoid slack space and allocating a single large buffer with efficient use of whatever buffer sizes the memory pool wants to give us. This allows us to use buffers that are sized appropriately for recycling instead of allocating enormous ones that must be thrown away.