Massive refactor around analyzer options
The goal is to eliminate all dependencies on MSBuild properties (which Unity cannot provide) or AdditionalFiles (which MSBuild and Unity support but are a pain to use).
Instead, we'll rely on a partial class with an attribute to trigger source generation. "custom types" that used to be specified via the AdditionalFiles will now be specified via assembly-targeted attributes.