Drop claim of support for VB and F#
The analyzers appear to support other languages, but not really, because in formulating diagnostics, they 'downcast' to C# syntax nodes to get good locations.
We can 'fix' this by making language-specific analyzer projects that share a lot of code, but unless we know there are VB/F# users out there, it's not worth it.