Guard against non-deterministic code generation
The problem was when more than one custom formatter is discovered for a given data type, a non-deterministic selection is made as to which formatter will be used by the generated resolver.
We fix this by producing an error in this condition, and suppressing generation of code that uses either formatter for the colliding data types.