Add private member serialization support for source generated formatters
As the formatter must have access (per C# rules) to the private members, the formatter must be nested under the data type.
This requires the data type (and any nesting types of that data type) to be declared as `partial` so that the source generated code can add the formatter as an `internal` member of it.
Since traditionally data types haven't been required to be partial, we only nest the formatter under the data type when private members must be serialized; otherwise we continue to generate the formatter as nested under the generated resolver.

Closes #1745