Add `CompositeResolverAttribute.IncludeLocalFormatters` property
This makes it much easier to get a combined resolver that includes all the formatters that are source generated or user authored, whether they format local data types or types from other assemblies.
Given the ability to also combine this resolver with a standard resolver, this allows the user to produce a fast, all-encompassing resolver for their application without listing even a single formatter.