Add MsgPack015 analyzer to encourage use of new `AllowPrivate` property
While source generators don't require it, it makes a real difference when using `DynamicObjectResolver` (which is the default, when a source generated formatter isn't found).

This analyzer comes with a code fix provider.

Plenty of tests (mostly testing source generated formatters) were using `internal` rather than `public` so with the analyzer enabled, a lot of test fixes were required.