Apply scoped in more places
More APIs that will benefit from having `scoped` applied to arguments.
These APIs fall into the pattern of `in / ref` parameter of
`ref struct` and return `ref struct`. The `scoped` prevents the `ref`
being calculated as part of the return lifetime. This makes the API
easier to use as it allows `in / ref` to stack bound values to be used
without that poisoning the return.